requirements.txt
Flask
cohere
numpy
python-dotenv
PyPDF2